changes99991

+++ api/models/basket.go
@@ -8,24 +8,24 @@ type Basket struct {
 	ID         string    `json:"id"`
 	Sale_id    string    `json:"sale_id"`
 	Product_id string    `json:"product_id"`
-	Quantity   float64   `json:"quantity"`
-	Price      float64   `json:"price"`
+	Quantity   uint64   `json:"quantity"`
+	Price      uint64   `json:"price"`
 	Create_at  time.Time `json:"ctreate_at"`
 }
 
 type CreateBasket struct {
 	Sale_id    string  `json:"sale_id"`
 	Product_id string  `json:"product_id"`
-	Quantity   float64 `json:"quantity"`
-	Price      float64 `json:"price"`
+	Quantity   uint64 `json:"quantity"`
+	Price      uint64 `json:"price"`
 }
 
 type UpdateBasket struct {
 	ID         string  `json:"id"`
 	Sale_id    string  `json:"sale_id"`
 	Product_id string  `json:"product_id"`
-	Quantity   float64 `json:"quantity"`
-	Price      float64 `json:"price"`
+	Quantity   uint64 `json:"quantity"`
+	Price      uint64 `json:"price"`
 }
 
 type BasketResponse struct {
