+++ api/models/product.go
@@ -7,7 +7,7 @@ import (
 type Product struct {
 	ID          string    `json:"id"`
 	Name        string    `json:"name"`
-	Price       float64   `json:"price"`
+	Price       uint64   `json:"price"`
 	Barcode     string    `json:"barcode"`
 	Category_id string    `json:"category_id"`
 	Create_at   time.Time `json:"create_at"`
@@ -15,7 +15,7 @@ type Product struct {
 
 type CreateProduct struct {
 	Name        string `json:"name"`
-	Price       float64   `json:"price"`
+	Price       uint64   `json:"price"`
 	Category_id string `json:"category_id"`
 	Barcode     int    `json:"barcode"`
 }
@@ -23,7 +23,7 @@ type CreateProduct struct {
 type UpdateProduct struct {
 	ID          string `json:"id"`
 	Name        string `json:"name"`
-	Price       float64   `json:"price"`
+	Price       uint64   `json:"price"`
 	Category_id string `json:"category_id"`
 }
