+++ api/models/staff.go
@@ -8,7 +8,7 @@ type Staff struct {
 	Tarif_id   string    `json:"tarif_id"`
 	Type_stuff string    `json:"type_stuff"`
 	Name       string    `json:"name"`
-	Balance    uint      `json:"balance"`
+	Balance    uint64    `json:"balance"`
 	Age        int       `json:"age"`
 	BirthDate  time.Time `json:"birthdate"`
 	Login      string    `json:"login"`
@@ -21,7 +21,7 @@ type CreateStaff struct {
 	Tarif_id   string `json:"tarif_id"`
 	Type_stuff string `json:"type_stuff"`
 	Name       string `json:"name"`
-	Balance    uint   `json:"balance"`
+	Balance    uint64 `json:"balance"`
 	BirthDate  string `json:"birthdate"`
 	Login      string `json:"login"`
 	Password   string `json:"password"`
@@ -33,7 +33,7 @@ type UpdateStaff struct {
 	TariffID  string `json:"tariff_id"`
 	StaffType string `json:"staff_type"`
 	Name      string `json:"name"`
-	Balance   uint   `json:"balance"`
+	Balance   uint64 `json:"balance"`
 	Login     string `json:"login"`
 }
 type StaffRepo struct {
