+++ api/models/storagetransaction.go
@@ -8,8 +8,8 @@ type TransactionStorage struct {
 	Staff_id         string    `json:"staff_id"`
 	Product_id       string    `json:"product_id"`
 	Transaction_type string    `json:"transaction_type"`
-	Price            float64   `json:"price"`
-	Quantity         float64   `json:"quantity"`
+	Price            uint64   `json:"price"`
+	Quantity         uint64   `json:"quantity"`
 	Create_at        time.Time `json:"ctreate_at"`
 }
 
@@ -18,8 +18,8 @@ type CreateTransactionStorage struct {
 	Staff_id         string  `json:"staff_id"`
 	Product_id       string  `json:"product_id"`
 	Transaction_type string  `json:"transaction_type"`
-	Price            float64 `json:"price"`
-	Quantity         float64 `json:"quantity"`
+	Price            uint64 `json:"price"`
+	Quantity         uint64 `json:"quantity"`
 }
 
 type UpdateTransactionStorage struct {
@@ -27,8 +27,8 @@ type UpdateTransactionStorage struct {
 	Branch_id  string  `json:"branch_id"`
 	Staff_id   string  `json:"staff_id"`
 	Product_id string  `json:"product_id"`
-	Price      float64 `json:"price"`
-	Quantity   float64 `json:"quantity"`
+	Price      uint64 `json:"price"`
+	Quantity   uint64 `json:"quantity"`
 }
 
 type TransactionStorageResponse struct {
